Finished global rafactoring (compiles)

+++ src/VM/Handlers/Opcode80D4Handler.cpp
@@ -47,7 +47,7 @@ void Opcode80D4Handler::_run()
     Logger::debug("SCRIPT") << "[80D4] [+] int objectPosition(GameObject* object)" << std::endl;
     auto object = _vm->dataStack()->popObject();
     if (!object) throw new Exception("Opcode 80d4 error");
-    _vm->dataStack()->push(object->hexagon()->number());
+    _vm->dataStack()->push((int)object->hexagon()->number());
 }
 
 }
